rename(): Register, Display 패키지 형성
