+++ src/main/java/lotto/controller/display/DisplayController.java
@@ -1,4 +1,4 @@
-package lotto.controller;
+package lotto.controller.display;
 
 public interface DisplayController {
     void process();
