+++ src/main/java/lotto/controller/register/RegisterBonusController.java
@@ -1,4 +1,4 @@
-package lotto.controller;
+package lotto.controller.register;
 
 import lotto.domain.Bonus;
 import lotto.dto.input.UserBonusDTO;
