+++ src/main/java/lotto/controller/register/RegisterController.java
@@ -1,4 +1,4 @@
-package lotto.controller;
+package lotto.controller.register;
 
 public interface RegisterController<E> {
     E process();
